0.5.0: Update dependencies
This was pulling in an outdated version of bluebird,
which couldn't be updated.

This resulted in confusing warnings when used with other libraries.

webpack was **not** updated because it defaults to production and
started minifying all files.
It's possible to set the environment to development when running
webpack, but I don't see how to do it in grunt
(didn't investigate deeply, low priority).